Make th evector array operators take a size_t idx.

+++ Pulp/Common/Math/XVector.h
@@ -63,12 +63,12 @@ public:
 		return *this;
 	}
 
-	X_INLINE const T& operator[](int i) const {
+	X_INLINE const T& operator[](size_t i) const {
 		X_ASSERT(i >= 0 && i < 2, "out of range")(i);
 		return (&x)[i]; 
 	}
 
-	X_INLINE T& operator[](int i) {
+	X_INLINE T& operator[](size_t i) {
 		X_ASSERT(i >= 0 && i < 2, "out of range")(i);
 			return (&x)[i];
 	}
@@ -286,12 +286,12 @@ public:
 		return *this;
 	}
 
-	X_INLINE const T& operator[](int i) const {
+	X_INLINE const T& operator[](size_t i) const {
 		X_ASSERT(i >= 0 && i < 3, "out of range")(i);
 			return (&x)[i];
 	}
 
-	X_INLINE T& operator[](int i) {
+	X_INLINE T& operator[](size_t i) {
 		X_ASSERT(i >= 0 && i < 3, "out of range")(i);
 			return (&x)[i];
 	}
@@ -615,12 +615,12 @@ public:
 		return *this;
 	}
 
-	X_INLINE const T& operator[](int i) const {
+	X_INLINE const T& operator[](size_t i) const {
 		X_ASSERT(i >= 0 && i < 4, "out of range")(i);
 			return (&x)[i];
 	}
 
-	X_INLINE T& operator[](int i) {
+	X_INLINE T& operator[](size_t i) {
 		X_ASSERT(i >= 0 && i < 4, "out of range")(i);
 			return (&x)[i];
 	}
@@ -892,12 +892,12 @@ public:
 		return *this;
 	}
 
-	X_INLINE const T& operator[](int i) const {
+	X_INLINE const T& operator[](size_t i) const {
 		X_ASSERT(i >= 0 && i < 5, "out of range")(i);
 		return (&x)[i];
 	}
 
-	X_INLINE T& operator[](int i) {
+	X_INLINE T& operator[](size_t i) {
 		X_ASSERT(i >= 0 && i < 5, "out of range")(i);
 		return (&x)[i];
 	}
